requirements.txt
streamlit==1.35.0
transformers==4.41.2
sentence-transformers==2.7.0
faiss-cpu==1.7.4
torch==2.3.0
accelerate==0.30.1
scipy==1.13.1
numpy==1.26.4
